Refactor userResolver and createUserCallback
